+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added address autocomplete via Gravity Forms country code/name data.
 - Added address autocomplete via Dutch PDOK.
 - Added address autocomplete via Google.
+- Added address autocomplete via Belgium local postal codes data.
 
 [Unreleased]: https://github.com/pronamic/wp-pronamic-maps/compare/1.0.1...HEAD
 [1.0.1]: https://github.com/pronamic/wp-pronamic-maps/compare/1.0.0...1.0.1